+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#4695](https://github.com/open-telemetry/opentelemetry-python/pull/4695)).
 - docs: linked the examples with their github source code location and added Prometheus example
   ([#4728](https://github.com/open-telemetry/opentelemetry-python/pull/4728))
+- semantic-conventions: Bump to 1.37.0
+  ([#4731](https://github.com/open-telemetry/opentelemetry-python/pull/4731))
 
 ## Version 1.36.0/0.57b0 (2025-07-29)